requirements.txt
Django==5.0
channels==4.0.0
daphne==4.0.0
google-generativeai
SpeechRecognition
gTTS
pydub
python-dotenv